% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_vignette}
\alias{eval_vignette}
\title{Determine if vignettes should be evaluated}
\usage{
eval_vignette()
}
\value{
Logical indicating whether to evaluate vignette code
}
\description{
Checks for vcr cassettes or API credentials to determine
if vignette code should be executed. Always returns FALSE during
R CMD check to avoid cassette mismatch errors.
}
\keyword{internal}
