% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{generate_grid_from_parameters}
\alias{generate_grid_from_parameters}
\title{Generate grid from parameter definition}
\usage{
generate_grid_from_parameters(parameters, control)
}
\description{
Generate grid from parameter definition
}
\keyword{internal}
