% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_ellmer_credentials}
\alias{has_ellmer_credentials}
\title{Check for ellmer credentials}
\usage{
has_ellmer_credentials()
}
\value{
Logical indicating if any LLM API keys are available
}
\description{
Check for ellmer credentials
}
\keyword{internal}
