% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{is_compiled}
\alias{is_compiled}
\title{Check if Module is Compiled}
\usage{
is_compiled(module)
}
\arguments{
\item{module}{A Predict module}
}
\value{
Logical indicating if the module is compiled
}
\description{
Check whether a module has been compiled/optimized.
}
