% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{map_inputs}
\alias{map_inputs}
\title{Map Inputs for Pipeline Steps}
\usage{
map_inputs(module, ...)
}
\arguments{
\item{module}{A Module object}

\item{...}{Named mappings: \code{output_field = "input_field"}}
}
\value{
A PipelineMappedModule object for use with \verb{\%>>\%}
}
\description{
Specifies how output fields from the previous step should be mapped to
input fields of this module. Use with \verb{\%>>\%} operator.
}
\examples{
\dontrun{
# Map 'documents' output to 'context' input
mod_retrieve \%>>\%
  map_inputs(mod_answer, documents = "context") \%>>\%
  mod_format
}
}
