% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{merge_optimization_control}
\alias{merge_optimization_control}
\title{Merge optimisation control defaults}
\usage{
merge_optimization_control(control)
}
\description{
Merge optimisation control defaults
}
\keyword{internal}
