% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{Predict}
\alias{Predict}
\title{Predict S7 Class (Internal)}
\usage{
Predict(signature = NULL, template = "", demos = list(), config = list())
}
\description{
Internal S7 class for prediction modules. Users should use the \code{module()}
function with \code{type = "predict"} instead of directly constructing Predict objects.
}
\seealso{
\code{\link[=module]{module()}} for the user-facing function
}
\keyword{internal}
