% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{prepare_candidate_grid}
\alias{prepare_candidate_grid}
\title{Prepare optimisation grid}
\usage{
prepare_candidate_grid(parameters, grid, control)
}
\description{
Prepare optimisation grid
}
\keyword{internal}
