% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleprompter-bootstrap.R
\name{print.BootstrapFewShot}
\alias{print.BootstrapFewShot}
\title{Print method for BootstrapFewShot}
\usage{
\method{print}{BootstrapFewShot}(x, ...)
}
\arguments{
\item{x}{A BootstrapFewShot object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for BootstrapFewShot
}
