% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleprompter-copro.R
\name{print.COPRO}
\alias{print.COPRO}
\title{Print method for COPRO}
\usage{
\method{print}{COPRO}(x, ...)
}
\arguments{
\item{x}{A COPRO object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for COPRO
}
