% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{print.dsprrr_batch_result}
\alias{print.dsprrr_batch_result}
\title{Print method for dsprrr_batch_result}
\usage{
\method{print}{dsprrr_batch_result}(x, ...)
}
\arguments{
\item{x}{A dsprrr_batch_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for dsprrr_batch_result
}
