% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{print.dsprrr_cost_summary}
\alias{print.dsprrr_cost_summary}
\title{Print method for dsprrr_cost_summary}
\usage{
\method{print}{dsprrr_cost_summary}(x, ...)
}
\arguments{
\item{x}{A dsprrr_cost_summary object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for dsprrr_cost_summary
}
