% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{print.dsprrr_evaluation}
\alias{print.dsprrr_evaluation}
\title{Print method for dsprrr_evaluation}
\usage{
\method{print}{dsprrr_evaluation}(x, ...)
}
\arguments{
\item{x}{A dsprrr_evaluation object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for dsprrr_evaluation
}
