% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer.R
\name{print.dsprrr_tool}
\alias{print.dsprrr_tool}
\title{Print method for dsprrr_tool}
\usage{
\method{print}{dsprrr_tool}(x, ...)
}
\arguments{
\item{x}{A dsprrr_tool object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for dsprrr_tool
}
