% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-core.R
\name{print.EvalResult}
\alias{print.EvalResult}
\title{Print method for EvalResult}
\usage{
\method{print}{EvalResult}(x, ...)
}
\arguments{
\item{x}{An EvalResult object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for EvalResult
}
