% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleprompter-gepa.R
\name{print.GEPA}
\alias{print.GEPA}
\title{Print method for GEPA}
\usage{
\method{print}{GEPA}(x, ...)
}
\arguments{
\item{x}{A GEPA object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for GEPA
}
