% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragnar.R
\name{print.ragnar_tool}
\alias{print.ragnar_tool}
\title{Print method for ragnar_tool}
\usage{
\method{print}{ragnar_tool}(x, ...)
}
\arguments{
\item{x}{A ragnar_tool object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for ragnar_tool
}
