% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleprompter-simba.R
\name{print.SIMBA}
\alias{print.SIMBA}
\title{Print method for SIMBA}
\usage{
\method{print}{SIMBA}(x, ...)
}
\arguments{
\item{x}{A SIMBA object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for SIMBA
}
