% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-logging.R
\name{print.Trial}
\alias{print.Trial}
\title{Print method for Trial}
\usage{
\method{print}{Trial}(x, ...)
}
\arguments{
\item{x}{A Trial object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for Trial
}
