% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-logging.R
\name{start_trial}
\alias{start_trial}
\title{Start a Trial}
\usage{
start_trial(trial)
}
\arguments{
\item{trial}{A Trial object.}
}
\value{
Updated Trial object with status "running".
}
\description{
Mark a trial as running and record the start time.
}
\keyword{internal}
