% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{with_inputs}
\alias{with_inputs}
\title{Inject Static Inputs for Pipeline Steps}
\usage{
with_inputs(module, ...)
}
\arguments{
\item{module}{A Module object}

\item{...}{Named values to inject as inputs}
}
\value{
A PipelineMappedModule object for use with \verb{\%>>\%}
}
\description{
Specifies constant values to inject as inputs to a module, regardless
of what the upstream module outputs.
}
\examples{
\dontrun{
# Inject a constant system prompt
mod_retrieve \%>>\%
  with_inputs(mod_answer, system_prompt = "Be very concise") \%>>\%
  mod_format
}
}
