% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compile S7 Generic and Methods}
\usage{
compile(teleprompter, program, ...)
}
\arguments{
\item{teleprompter}{A Teleprompter object}

\item{program}{A module to optimize}

\item{...}{Additional arguments}

\item{trainset}{Training data}
}
\value{
An optimized module
}
\description{
Generic method for compiling/optimizing a module using a teleprompter.
}
\details{
This file defines the compile generic and its methods for optimizing
DSPrrr modules using teleprompters.
Compile Generic
}
