% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsprrr-package.R
\docType{package}
\name{dsprrr-package}
\alias{dsprrr}
\alias{dsprrr-package}
\title{dsprrr: Declarative Self-Improving Language Programs}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Build declarative, test-driven, and optimizable applications with large language models (LLMs). Implements ideas from the 'DSPy' framework described by Khattab et al. (2023) \doi{10.48550/arXiv.2310.03714} through typed signatures, composable modules, evaluation, and data-driven prompt optimization. Integrates with 'ellmer' for model access and supports tracing, persistence, caching, and asynchronous execution. Designed for systematic improvement of language-model workflows in 'R'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JamesHWade/dsprrr}
  \item \url{https://jameshwade.github.io/dsprrr/}
  \item Report bugs at \url{https://github.com/JamesHWade/dsprrr/issues}
}

}
\author{
\strong{Maintainer}: James Wade \email{github@jameshwade.com} (\href{https://orcid.org/0000-0002-9740-1905}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item James Wade \email{github@jameshwade.com} (\href{https://orcid.org/0000-0002-9740-1905}{ORCID}) [copyright holder]
  \item Anatoliy Sokolov \email{ansokolov@dow.com}
}

}
\keyword{internal}
