% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsprrr-package.R
\docType{package}
\name{dsprrr-package}
\alias{dsprrr}
\alias{dsprrr-package}
\title{dsprrr: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JamesHWade/dsprrr}
  \item Report bugs at \url{https://github.com/JamesHWade/dsprrr/issues}
}

}
\author{
\strong{Maintainer}: James Wade \email{github@jameshwade.com} (\href{https://orcid.org/0000-0002-9740-1905}{ORCID})

}
\keyword{internal}
