% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleprompter.R
\name{GridSearchTeleprompter}
\alias{GridSearchTeleprompter}
\title{GridSearchTeleprompter}
\usage{
GridSearchTeleprompter(
  metric = NULL,
  metric_threshold = NULL,
  max_errors = 5L,
  variants = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  k = 2L,
  eval_sample_size = 50L,
  verbose = TRUE
)
}
\description{
A teleprompter that performs grid search over different instruction
and template variants to find the best performing configuration.
}
