% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input_helpers}
\alias{input_helpers}
\alias{input_string}
\alias{input_number}
\alias{input_boolean}
\alias{input_integer}
\alias{input_enum}
\alias{input_array}
\alias{input_object}
\title{Create typed input helpers for common cases}
\usage{
input_string(name, description = NULL, ...)

input_number(name, description = NULL, ...)

input_boolean(name, description = NULL, ...)

input_integer(name, description = NULL, ...)

input_enum(name, values, description = NULL, ...)

input_array(name, item_type = ellmer::type_string(), description = NULL, ...)

input_object(name, ..., description = NULL)
}
\arguments{
\item{name}{Name of the input field}

\item{description}{Optional description of the input}

\item{...}{Additional arguments passed to the type constructor}

\item{values}{For input_enum, the allowed values}

\item{item_type}{For input_array, the type of array items}
}
\description{
Create typed input helpers for common cases
}
