% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleprompter.R
\name{LabeledFewShot}
\alias{LabeledFewShot}
\title{LabeledFewShot Teleprompter}
\usage{
LabeledFewShot(
  metric = NULL,
  metric_threshold = NULL,
  max_errors = 5L,
  k = 4L,
  sample = TRUE,
  seed = 123L
)
}
\description{
A simple teleprompter that adds labeled examples from the training set
as demonstrations to the module. This is the simplest form of few-shot
learning.
}
