% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-logging.R
\name{load_trial_log}
\alias{load_trial_log}
\title{Load Trial Log from Directory}
\usage{
load_trial_log(log_dir)
}
\arguments{
\item{log_dir}{Path to the log directory.}
}
\value{
A TrialLog object.
}
\description{
Load a TrialLog from a directory that was previously saved.
}
\examples{
\dontrun{
log <- load_trial_log("logs/my_optimizer/")
log$as_tibble()
}
}
