% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleprompter-better-together.R
\name{print.BetterTogether}
\alias{print.BetterTogether}
\title{Print method for BetterTogether}
\usage{
\method{print}{BetterTogether}(x, ...)
}
\arguments{
\item{x}{A BetterTogether object}

\item{...}{Additional arguments}
}
\description{
Print method for BetterTogether
}
\keyword{internal}
