% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer-logging.R
\name{read_trials_jsonl}
\alias{read_trials_jsonl}
\title{Read Trials from JSONL File}
\usage{
read_trials_jsonl(path)
}
\arguments{
\item{path}{File path to the JSONL file.}
}
\value{
A list of Trial objects.
}
\description{
Read Trial objects from a JSONL file.
}
\examples{
\dontrun{
trials <- read_trials_jsonl("trials.jsonl")
}
}
