% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip.R
\name{temperature}
\alias{temperature}
\title{Temperature Parameter for dials}
\usage{
temperature(range = c(0, 1), trans = NULL)
}
\arguments{
\item{range}{Range of temperature values (default c(0, 1)).}

\item{trans}{Transformation (default NULL for identity).}
}
\value{
A dials parameter object.
}
\description{
Creates a dials parameter object for LLM temperature.
}
\examples{
\dontrun{
temperature()
temperature(range = c(0.1, 0.9))
}
}
